ajout de la getion des erreurs 404 et 500

+++ front/front-suppliers/app.py
@@ -12,6 +12,14 @@ gateway_url = '10.226.159.191:9090'
 products_service = ProductService(gateway_url)
 app.secret_key = "this_is_a_secret_key_get_over_it"
 
+@app.errorhandler(404)
+def not_found(e):
+	return render_template("notFound.html"), 404
+
+@app.errorhandler(500)
+def everything_is_broken(e):
+	return render_template("serverError.html"), 500
+
 @app.route("/")
 def root():
 
@@ -47,19 +55,26 @@ def logout():
 
 @app.route("/products")
 def products_list():
-    products = products_service.retrieve_products(get_id_supplier())
-    categories = products_service.retrieve_categories()
-    session['active'] = 'products'
-    return render_template("products/products.html", products=products, categories=categories)
+
+	 if "logged" not in session or session["logged"] == "FALSE" :
+		 return redirect(url_for("root"))
+		 products = products_service.retrieve_products(get_id_supplier())
+		 categories = products_service.retrieve_categories()
+		 session['active'] = 'products'
+		 return render_template("products/products.html", products=products, categories=categories)
 
 @app.route("/addproduct")
 def add_product():
+	if "logged" not in session or session["logged"] == "FALSE" :
+		return redirect(url_for("root"))
 	session['active'] = 'add'
 	categories = products_service.retrieve_categories()
 	return render_template("products/addproduct.html", categories=categories)
 
 @app.route("/updateinfo")
 def update_info():
+	if "logged" not in session or session["logged"] == "FALSE" :
+		return redirect(url_for("root"))
 	session['active'] = 'updateinfo'
 	ID = session["supplier"]['id']
 	supplier = supplier_service.get_supplier(ID)
@@ -67,6 +82,8 @@ def update_info():
 
 @app.route("/supplierUdapte", methods=["POST"])
 def supplier_update():
+	if "logged" not in session or session["logged"] == "FALSE" :
+		 	return redirect(url_for("root"))
 	ID = session["supplier"]["id"]
 	dic = {}
 	for key in request.form.keys():
@@ -77,6 +94,8 @@ def supplier_update():
 
 @app.route("/createProduct", methods=['POST'])
 def create_product():
+	if "logged" not in session or session["logged"] == "FALSE" :
+		return redirect(url_for("root"))
 	product = Product(id='', id_supplier=get_id_supplier(), **request.form)
 	res = products_service.create_product(product)
 	if res == 'OK':
@@ -87,12 +106,16 @@ def create_product():
 
 @app.route("/updateProduct/<id>", methods=['POST'])
 def update_product(id):
-    product = Product(id, id_supplier=get_id_supplier(), **request.form)
-    products_service.update_product(product)
-    return render_template("products/success.html", previous_page="/products", message="Produit mis à jour avec succès")
+	if "logged" not in session or session["logged"] == "FALSE" :
+		return redirect(url_for("root"))
+		product = Product(id, id_supplier=get_id_supplier(), **request.form)
+		products_service.update_product(product)
+		return render_template("products/success.html", previous_page="/products", message="Produit mis à jour avec succès")
 
 @app.route("/deleteProduct/<id>", methods=['DELETE'])
 def delete_product(id):
+	if "logged" not in session or session["logged"] == "FALSE" :
+			return redirect(url_for("root"))
 	products_service.delete_product(id)
 	return "OK"
